[IDLE-104] 바텀시트 ViewController컴포넌트 구현
